requirements.txt
pandas
numpy
scikit-learn
pyyaml
matplotlib
seaborn
joblib
